Fixed ALL CAPS text bug in wxX11 Removed compiler warning in text sample
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_2_4_BRANCH@17922 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ samples/text/text.cpp
@@ -1183,7 +1183,7 @@ void MyFrame::OnFileSave(wxCommandEvent& event)
         wxLogStatus(this,
                     _T("Successfully saved file (text len = %lu, file size = %ld)"),
                     (unsigned long)m_panel->m_textrich->GetValue().length(),
-                    file.Length());
+                    (long) file.Length());
 #endif
     }
     else
